Initial development
MAC can take two arrays and perform multiply accumulate operation after
a number of clock cycles that is equal to the size of the input array.

Currently, I assume that the user (well... me) will supply two flattened
arrays that are equal in size, and that the user will not change the
input array during the multiply operation. (Maybe I should add a lock
flag to the MAC mocule).

Test_wrapper is intended to be a testbench for MAC (and FIR later)